+++ src/element/create.js
@@ -10,7 +10,7 @@ import set from './set'
 import createState from './state'
 import update from './update'
 import * as on from '../event/on'
-import { assignClass } from './params/classList'
+import { assignClass } from './mixins/classList'
 import { isFunction, isNumber, isString } from '../utils'
 import { remove, lookup, log, keys, parse } from './methods'
 // import { overwrite, clone, fillTheRest } from '../utils'
