+++ src/element/mixins/content.js
