+++ src/element/mixins/registry.js
