+++ src/element/mixins/text.js
